完成tank类基础，显示player1并能移动和墙体碰撞，暂无其它碰撞检测

+++ FC经典坦克/Font.cpp
@@ -98,7 +98,8 @@ void Font::show(int x, int y, int interval) {
 		//EasyXÑùÊ½ÉèÖÃ
 		settextcolor(color);
 		settextstyle(&f);
-		outtextxy((int)x, (int)y, s);
+
+		outtextxy(x, y, s);
 	}
 }
